samples/python/multiprocess_detection/requirements.txt
flask==1.1.4
cython
scipy
numpy
pillow
sklearn
requests==2.22.0
configobj==5.0.6
pyDes==2.0.1
PyMySQL==0.9.3
borax==3.2.0
tornado==5.1
redis==3.2.1
psutil==5.7.0